Update drag scroll from ploopy PR

+++ drag_scroll/drag_scroll.h
@@ -4,11 +4,13 @@
 #pragma once
 
 #include <stdlib.h>
+#include <stdint.h>
+#include <stdbool.h>
 
-float get_drag_scroll_h_divisor(void);
-float get_drag_scroll_v_divisor(void);
-void  set_drag_scroll_h_divisor(float divisor);
-void  set_drag_scroll_v_divisor(float divisor);
-void  set_drag_scroll_divisor(float divisor);
-bool  get_drag_scroll_scrolling(void);
-void  set_drag_scroll_scrolling(bool scrolling);
+int8_t get_drag_scroll_h_divisor(void);
+int8_t get_drag_scroll_v_divisor(void);
+void   set_drag_scroll_h_divisor(int8_t divisor);
+void   set_drag_scroll_v_divisor(int8_t divisor);
+void   set_drag_scroll_divisor(int8_t divisor);
+bool   get_drag_scroll_scrolling(void);
+void   set_drag_scroll_scrolling(bool scrolling);
